% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_probs.R
\name{joint_probs}
\alias{joint_probs}
\title{Calculate all joint probabilities from a pair of ancestral state
reconstructions}
\usage{
joint_probs(x, y, include.tips = TRUE)
}
\arguments{
\item{x, }{y marginal probabilities; generally the \code{states} element from
a \code{rayDISC} object.}

\item{include.tips}{[optional] logical indicating whether or not joint
probabilities for terminal nodes should be included in output. Default is
\code{TRUE}.}
}
\description{
Calculate all joint probabilities from a pair of ancestral state
reconstructions
}
\references{
a list with a \code{states} matrix of joint probabilities for
ancestral state combinations at internal nodes and if \code{include.tips =
TRUE}, a \code{tip.states} matrix of joint probabilities of the state
combination at the terminal nodes.
}
